+++ generic-override/src/Data/Override/Instances.hs
@@ -7,11 +7,9 @@
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 module Data.Override.Instances () where
 
-import Data.Coerce (Coercible, coerce)
 import Data.Function (on)
+import Data.Override.Internal (Override)
 import GHC.Generics (Generic(Rep, from, to))
-
-import Data.Override.Internal
 
 -- The @foo `on` from'@ idiom is taken from @generic-data@ by Li-yao Xia.
 from' :: Generic a => a -> Rep a ()
@@ -29,13 +27,6 @@ instance
   where
   (==) = (==) `on` from'
 
-instance
-  ( Coercible a (Using ms a xs)
-  , Eq (Using ms a xs)
-  ) => Eq (Overridden ms a xs)
-  where
-  x == y = (==) @(Using ms a xs) (coerce x) (coerce y)
-
 -- Ord
 
 instance
@@ -44,14 +35,6 @@ instance
   ) => Ord (Override a xs)
   where
   compare = compare `on` from'
-
-instance
-  ( Coercible a (Using ms a xs)
-  , Ord (Using ms a xs)
-  ) => Ord (Overridden ms a xs)
-  where
-  compare x y = compare @(Using ms a xs) (coerce x) (coerce y)
-
 
 -- Semigroup
 
@@ -62,13 +45,6 @@ instance
   where
   x <> y = to (from' x <> from' y)
 
-instance
-  ( Coercible a (Using ms a xs)
-  , Semigroup (Using ms a xs)
-  ) => Semigroup (Overridden ms a xs)
-  where
-  x <> y = coerce $ (<>) @(Using ms a xs) (coerce x) (coerce y)
-
 -- Monoid
 
 instance
@@ -78,11 +54,3 @@ instance
   where
   mempty = to' mempty
   x `mappend` y = to (from' x `mappend` from' y)
-
-instance
-  ( Coercible a (Using ms a xs)
-  , Monoid (Using ms a xs)
-  ) => Monoid (Overridden ms a xs)
-  where
-  mempty = coerce $ mempty @(Using ms a xs)
-  x `mappend` y = coerce $ mappend @(Using ms a xs) (coerce x) (coerce y)