+++ generic-override/test/Encode.hs
@@ -1,8 +1,10 @@
 {-# LANGUAGE DefaultSignatures #-}
+{-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
@@ -10,9 +12,8 @@
 {-# LANGUAGE UndecidableInstances #-}
 module Encode where
 
-import Data.Coerce
-import Data.List
-import Data.Override.Internal
+import Data.List (intercalate)
+import Data.Override (Override)
 import GHC.Generics
 
 class Encode a where
@@ -21,18 +22,9 @@ class Encode a where
   encode = gencode . from
 
 instance
-  ( Generic a
-  , GOverride xs (Rep a)
+  ( Generic (Override a xs)
   , GEncode (Rep (Override a xs))
   ) => Encode (Override a xs)
-
-instance
-  ( u ~ Using ms a xs
-  , Coercible a u
-  , Encode u
-  ) => Encode (Overridden ms a xs)
-  where
-  encode = encode @u . coerce
 
 class GEncode f where
   gencode :: f a -> String
@@ -70,11 +62,12 @@ instance Encode Int where
 instance Encode Char where
   encode = pure
 
+-- | Create an overlapping instance to verify that overriding with
+-- 'ListOf' avoids this instance.
 instance {-# OVERLAPPING #-} Encode String where
   encode = id
 
-instance (Encode a) => Encode [a] where
-  encode = intercalate "," . map encode
+deriving via (ListOf a) instance (Encode a) => Encode [a]
 
 newtype ListOf a = ListOf { unListOf :: [a] }